core/graph_store.py
"""
Nebula Graph Store Implementation
==================================

Implements GraphStoreInterface for NebulaGraph.
"""

from typing import List, Dict, Any, Optional
import time
from core.service_interfaces import GraphStoreInterface, Relationship
from core.logging_config import Logger, log_performance


logger = Logger(__name__)


class NebulaGraphStore(GraphStoreInterface):
    """
    Production Nebula graph store implementation.
    Centralizes all graph operations with proper error handling.
    """

    def __init__(
        self,
        host: str = "127.0.0.1",
        port: int = 9669,
        user: str = "root",
        password: str = "nebula",
        space: str = "extraction_db",
    ):
        self.host = host
        self.port = port
        self.user = user
        self.password = password
        self.space = space
        self.pool = None
        self.initialized = False

    @log_performance
    def initialize(self) -> bool:
        """Initialize Nebula connection and create schema"""
        try:
            from nebula3.gclient.net import ConnectionPool
            from nebula3.Config import Config

            config = Config()
            config.max_connection_pool_size = 10
            config.timeout = 30000

            self.pool = ConnectionPool()
            ok = self.pool.init([(self.host, self.port)], config)

            if not ok:
                logger.error("Failed to initialize Nebula connection pool")
                return False

            logger.info("✓ Nebula connection pool initialized")

            # Create space and schema
            self._create_space_and_schema()
            self.initialized = True
            return True

        except ImportError:
            logger.error("nebula3-python package not installed")
            return False
        except Exception as e:
            logger.error(f"Nebula initialization failed: {e}", exc_info=True)
            return False

    def _get_session(self):
        """Get authenticated session"""
        if not self.pool:
            return None
        try:
            return self.pool.get_session(self.user, self.password)
        except Exception as e:
            logger.error(f"Failed to get session: {e}")
            return None

    def _create_space_and_schema(self):
        """Create space and entity/edge schema"""
        session = self._get_session()
        if not session:
            logger.error("Cannot create schema - no session")
            return

        try:
            # Create space
            session.execute(f"""
                CREATE SPACE IF NOT EXISTS {self.space}(
                    vid_type=FIXED_STRING(256)
                )
            """)

            time.sleep(2)
            session.execute(f"USE {self.space}")

            # Create entity tags
            tags = {
                "Person": "name string, confidence float",
                "Organization": "name string, confidence float",
                "Date": "value string, confidence float",
                "Amount": "value string, confidence float",
                "Location": "name string, confidence float",
                "Project": "name string, confidence float",
                "Invoice": "number string, confidence float",
                "Agreement": "title string, confidence float",
                "Shipment": "id string, confidence float",
                "Port": "name string, confidence float",
                "Supplier": "name string, confidence float",
                "Incident": "title string, confidence float",
            }

            for tag_name, properties in tags.items():
                session.execute(f"""
                    CREATE TAG IF NOT EXISTS `{tag_name}`(
                        {properties}
                    )
                """)

            # Create edge types for relationships
            edges = [
                "WORKS_AT", "EMPLOYS", "LOCATED_IN", "HOSTS",
                "WORKS_ON", "MANAGES", "MANAGED_BY", "PARTY_TO",
                "INVOLVED_IN", "HAS_VALUE", "SIGNED_ON", "ISSUED",
                "ISSUED_BY", "BILLED_TO", "SENT", "SENT_BY",
                "RECEIVED_BY", "ROUTED_THROUGH", "AFFECTED",
                "EXPERIENCED", "CAUSED_BY", "RELATED_TO",
            ]

            for edge in edges:
                session.execute(f"""
                    CREATE EDGE IF NOT EXISTS `{edge}`(
                        confidence float,
                        context string
                    )
                """)

            logger.info(f"✓ Nebula schema created for space '{self.space}'")

        except Exception as e:
            logger.error(f"Schema creation error: {e}")
        finally:
            session.release() if session else None

    @log_performance
    def store_entities(self, entities: List[Dict[str, Any]]) -> Dict[str, int]:
        """Store entities in graph"""
        if not self.initialized:
            logger.error("Graph store not initialized")
            return {}

        session = self._get_session()
        if not session:
            logger.error("Cannot get session for entity storage")
            return {}

        result = {}

        try:
            session.execute(f"USE {self.space}")

            for entity in entities:
                entity_type = entity.get("type", "").title()
                entity_value = entity.get("value", "")
                confidence = entity.get("confidence", 0.8)

                if not entity_type or not entity_value:
                    continue

                # Sanitize values
                entity_value = entity_value.replace("'", "\\'")

                try:
                    query = f"""
                        INSERT VERTEX `{entity_type}`(name, confidence)
                        VALUES '{entity_value}':('{entity_value}', {confidence})
                    """
                    session.execute(query)

                    if entity_type not in result:
                        result[entity_type] = 0
                    result[entity_type] += 1

                except Exception as e:
                    logger.debug(f"Entity insert error: {e}")

            logger.info(f"Stored entities: {result}")
            return result

        except Exception as e:
            logger.error(f"Entity storage failed: {e}", exc_info=True)
            return result
        finally:
            session.release() if session else None

    @log_performance
    def store_relationships(self, relationships: List[Relationship]) -> int:
        """Store relationships in graph"""
        if not self.initialized:
            logger.error("Graph store not initialized")
            return 0

        session = self._get_session()
        if not session:
            logger.error("Cannot get session for relationship storage")
            return 0

        count = 0

        try:
            session.execute(f"USE {self.space}")

            for rel in relationships:
                try:
                    source = rel.source.replace("'", "\\'")
                    target = rel.target.replace("'", "\\'")
                    rel_type = rel.relationship_type.upper()
                    confidence = rel.confidence
                    context = (rel.context or "").replace("'", "\\'")[:255]

                    query = f"""
                        INSERT EDGE `{rel_type}`(confidence, context)
                        VALUES '{source}' -> '{target}':(
                            {confidence},
                            '{context}'
                        )
                    """
                    session.execute(query)
                    count += 1

                except Exception as e:
                    logger.debug(f"Relationship insert error: {e}")

            logger.info(f"Stored {count} relationships")
            return count

        except Exception as e:
            logger.error(f"Relationship storage failed: {e}", exc_info=True)
            return count
        finally:
            session.release() if session else None

    @log_performance
    def store_graph(
        self,
        entities: List[Dict[str, Any]],
        relationships: List[Relationship],
    ) -> Dict[str, Any]:
        """Store entire graph atomically"""
        result = {
            "entities_added": 0,
            "relationships_added": 0,
        }

        try:
            # Store entities
            entity_result = self.store_entities(entities)
            result["entities_added"] = sum(entity_result.values())

            # Store relationships
            result["relationships_added"] = self.store_relationships(relationships)

            logger.info(
                f"Graph storage complete: "
                f"{result['entities_added']} entities, "
                f"{result['relationships_added']} relationships"
            )

            return result

        except Exception as e:
            logger.error(f"Graph storage failed: {e}", exc_info=True)
            return result

    @log_performance
    def query_relationships(
        self,
        entity_name: str,
        entity_type: Optional[str] = None,
        relationship_type: Optional[str] = None,
    ) -> List[Relationship]:
        """Query relationships for an entity"""
        if not self.initialized:
            logger.error("Graph store not initialized")
            return []

        session = self._get_session()
        if not session:
            logger.error("Cannot get session for relationship query")
            return []

        relationships = []

        try:
            session.execute(f"USE {self.space}")

            entity_name = entity_name.replace("'", "\\'")

            # Multi-hop traversal
            query = f"""
                FETCH PROP ON * '{entity_name}' | 
                GO 1 TO 2 STEPS FROM $-.id OVER *
                YIELD $-.id AS source, $$.id AS target, edge as e
            """

            result = session.execute(query)

            # Process results into Relationship objects
            # This is a simplified version - actual parsing depends on result format
            logger.debug(f"Relationship query returned results for {entity_name}")

            return relationships

        except Exception as e:
            logger.debug(f"Relationship query error: {e}")
            return relationships
        finally:
            session.release() if session else None

    @log_performance
    def find_paths(
        self,
        source: str,
        target: str,
        max_depth: int = 3,
    ) -> List[List[Relationship]]:
        """Find paths between entities"""
        if not self.initialized:
            return []

        session = self._get_session()
        if not session:
            return []

        paths = []

        try:
            session.execute(f"USE {self.space}")

            source = source.replace("'", "\\'")
            target = target.replace("'", "\\'")

            # Find shortest path
            query = f"""
                FIND SHORTEST PATH FROM '{source}' TO '{target}'
                OVER * REVERSIBLE UPTO {max_depth} STEPS
                YIELD path as p
            """

            result = session.execute(query)
            # Parse paths from result
            logger.debug(f"Found paths from {source} to {target}")

            return paths

        except Exception as e:
            logger.debug(f"Path finding error: {e}")
            return paths
        finally:
            session.release() if session else None
